+++ acp/acp-spring-boot-starter/src/main/java/pers/acp/springboot/core/exceptions/ServerException.java
@@ -1,10 +1,15 @@
 package pers.acp.springboot.core.exceptions;
 
 import pers.acp.core.exceptions.base.BaseException;
+import pers.acp.springboot.core.enums.ResponseCode;
 
 public class ServerException extends BaseException {
 
     private static final long serialVersionUID = -826325914171274124L;
+
+    public ServerException(ResponseCode responseCode) {
+        super(responseCode.getValue(), responseCode.getName());
+    }
 
     public ServerException(String message) {
         super(message);