+++ doc/version_history.md
@@ -8,7 +8,8 @@
     6. 更新各依赖包版本
     7. gateway 增加自定义断路器
     8. 修改 Eureka Server 及 Client 配置，优化服务注册/发现感知度
-    9. 优化 Controller 切片日志
+    9. 优化 Controller 切片日志，重命名切片类
+    10. kotlin 版本升级至 1.2.40，增加 kotlin demo，test:testkotlin
 ##### 3.9.0
     1. 升级 spring-cloud 版本为 Finchley.M9
     2. spring cloud 网关服务由 zuul 更换为 spring-cloud-gateway